el test de digest falla poqrue TextEncoder no está implementado en jsdom
