+++ jest.config.js
@@ -140,8 +140,9 @@ module.exports = {
     // snapshotSerializers: [],
 
     // The test environment that will be used for testing
-    // testEnvironment: "jest-environment-jsdom",
+    // testEnvironment: 'jest-environment-jsdom',
     // testEnvironment: 'node',
+    // testEnvironment: '<rootDir>/__tests__/env/custom-test-env.js',
 
     // Options that will be passed to the testEnvironment
     // testEnvironmentOptions: {},
